requirements-linux.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.3.0
bitarray==3.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
ckzg==2.1.1
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
cytoolz==1.0.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
Envelopes==0.4
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
executing==2.2.0
fonttools==4.58.5
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.3.4
openai==1.99.9
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyunormalize==16.0.0
pyzbar==0.1.9
pyzmq==26.4.0
qrcode==8.2
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rlp==4.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.22
wcwidth==0.2.13
web3==7.12.1
websockets==15.0.1
whitenoise==6.9.0
yarl==1.20.1
